nmci: add option to synchronously reload in nmutil.reload_NM_service()
Sending a SIGHUP signal is asynchronous (meaning, after the signal was
sent, we don't know whether the change was already processed by
NetworkManager). For that there is a D-Bus call, which is supposed to
only return after the changes where done. We could either make that call
via busctl, or call `systemctl reload NetworkManager`. The effect is
basically the same, because the systemd reload just does the same busctl
call (see NetworkManager.service file).

+++ nmci/nmutil.py
@@ -108,8 +108,14 @@ class _NMUtil:
         print("reload NM connections")
         nmci.process.nmcli("con reload")
 
-    def reload_NM_service(self):
+    def reload_NM_service(self, synchronous=False):
         print("reload NM service")
+        if synchronous:
+            # ExecReload= uses busctl and waits for a response.
+            r = nmci.process.systemctl("reload NetworkManager.service")
+            assert r.returncode == 0, f"systemctl reload NetworkManager failed with {r}"
+            return
+        # Send an async SIGHUP signal.
         nmci.process.run_stdout("pkill -HUP NetworkManager")
         timeout = nmci.util.start_timeout(self.DEFAULT_TIMEOUT)
         self.wait_for_nm_bus(timeout)
